pid autotuner class
